Feat(room): Update FindRoomCommand to also search for matching tags
